winkernel: add proxier_sync
